Dealer and Deck seperate
Deck is now it's own class to allow for ease of accesing deck structure
