+++ poker/dealer.py
@@ -1,9 +1,9 @@
-from random import shuffle
+from deck import Deck
 
 
 class Dealer:
     def __init__(self):
-        self.deck = self.init_deck()
+        self.deck = Deck.init_deck()
 
     def deal(self, n_cards):
         return [self.deck.pop(0) for _ in range(n_cards)]
@@ -13,11 +13,3 @@ class Dealer:
 
     def cards_left(self):
         return len(self.deck)
-
-    @classmethod
-    def init_deck(cls):
-        suits = ["d", "h", "s", "c"] # d: DIAMOND, h: HEART, s: SPADE, c: CLUB
-        ranks = ["2", "3", "4", "5", "6", "7", "8", "9", "10", "J", "Q", "K", "A"]
-        deck = [rank+suit for rank in ranks for suit in suits]
-        shuffle(deck)
-        return deck
